+++ vm/src/platform/process_memory_semantics.h
@@ -53,8 +53,8 @@ public:
 private:
   static char  mmap_filename[BUFSIZ];
 public:
-  static char* map_heap_memory(int total_file_size, int bytes_to_map,
-                               int page_size_used_in_heap_arg, void* where, int offset,
+  static char* map_heap_memory(off_t total_file_size, size_t bytes_to_map,
+                               size_t page_size_used_in_heap_arg, void* where, off_t offset,
                                int main_pid, int flags);
   
 };
